+++ src/main/java/ladder/view/output/ConsoleOutputView.java
@@ -1,4 +1,4 @@
-package ladder.view;
+package ladder.view.output;
 
 import ladder.domain.Result;
 import ladder.domain.Rewards;
@@ -7,16 +7,16 @@ import ladder.domain.person.People;
 import ladder.util.LadderOutputConverter;
 
 
-public class OutputView {
+public class ConsoleOutputView implements OutputView{
 
-    public static void ladderCreateResult(People people, LadderLines ladderLines, Rewards rewards) {
+    public void ladderCreateResult(People people, LadderLines ladderLines, Rewards rewards) {
         System.out.println("사다리 결과");
         System.out.println(LadderOutputConverter.peopleOutput(people));
         System.out.println(LadderOutputConverter.ladderLinesOutput(ladderLines));
         System.out.println(LadderOutputConverter.rewardsOutput(rewards));
     }
 
-    public static void gameResult(Result result) {
+    public void gameResult(Result result) {
         System.out.println("실행 결과");
         System.out.println(LadderOutputConverter.resultOutput(result));
     }
